🔧 chore: 절대 경로 사용, constants 데이터 적합한 파일로 이동 #24
