+++ src/pages/ApplicationDocuments/ApplicationDocumentsPage.tsx
@@ -19,12 +19,6 @@ const mockDocumentsSummaryResponse: DocumentsSummaryResponse = {
       word_url: "https://example.com/contracts/standard_2001.docx",
       status: "BEFORE_CONFIRMATION"
     },
-    integrated_application: {
-      id: 3001,
-      pdf_url: "https://example.com/applications/integrated_3001.pdf",
-      hwp_url: "https://example.com/applications/integrated_3001.hwp",
-      word_url: "https://example.com/applications/integrated_3001.docx"
-    }
 };
 
 const ApplicationDocumentsPage = () => {
