+++ src/router.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import HomePage from '@/pages/Home/HomePage';
-import InformationPage from './pages/Information/InformationPage';
-import ApplicationDocumentsPage from './pages/ApplicationDocuments/ApplicationDocumentsPage';
+import InformationPage from '@/pages/Information/InformationPage';
+import ApplicationDocumentsPage from '@/pages/ApplicationDocuments/ApplicationDocumentsPage';
 
 const Router = () => {
   return (
