+++ src/types/api/document.ts
@@ -31,19 +31,3 @@ export enum DocumentType {
   LABOR_CONTRACT = 'standard_labor_contract',
   INTEGRATED_APPLICATION = 'integrated_application',
 }
-
-// 문서 타입별 정보를 담은 객체
-export const DocumentTypeInfo = {
-  [DocumentType.PART_TIME_PERMIT]: {
-    name: 'Part-Time Work Permit Form',
-    key: 'part_time_employment_permits',
-  },
-  [DocumentType.LABOR_CONTRACT]: {
-    name: 'Employment Contract',
-    key: 'standard_labor_contract',
-  },
-  [DocumentType.INTEGRATED_APPLICATION]: {
-    name: 'Integrated Application Form',
-    key: 'integrated_application',
-  },
-} as const;
